+++ backend/tests/Functional/Controller/User/AuthControllerTest.php
@@ -210,7 +210,7 @@ class AuthControllerTest extends WebTestCase
     {
         $h = new Helper();
         $h->emptyDb();
-        $h->addRoles([Role::USER, Role::TRACKING, Role::WATCHLIST, Role::WATCHLIST_MANAGER]);
+        $h->addRoles([Role::USER, Role::TRACKING, Role::WATCHLIST, Role::WATCHLIST_MANAGER, Role::GROUP_MANAGER]);
 
         list($token, $keySet) = Helper::generateToken(['read-this', 'and-this']);
         $state = '1jdHR64hSdYf';
@@ -284,7 +284,7 @@ class AuthControllerTest extends WebTestCase
     {
         $h = new Helper();
         $h->emptyDb();
-        $h->addRoles([Role::TRACKING, Role::WATCHLIST, Role::WATCHLIST_MANAGER]);
+        $h->addRoles([Role::TRACKING, Role::WATCHLIST, Role::WATCHLIST_MANAGER, Role::GROUP_MANAGER]);
 
         $h->addCharacterMain('User1', 654, [Role::USER], ['group1']);
         $this->loginUser(654);